// types/event.ts
export type CreateEventParams = {
    userId: string;
    event: {
        title: string;
        description: string;
        location: string;
        imageUrl: string;
        startDateTime: Date;
        endDateTime: Date;
        categoryId: string;
        price: string;
        isFree: boolean;
        url: string;
    };
    path: string;
};

export type UpdateEventParams = {
    userId: string;
    event: {
        _id: string;
        title: string;
        imageUrl: string;
        description: string;
        location: string;
        startDateTime: Date;
        endDateTime: Date;
        categoryId: string;
        price: string;
        isFree: boolean;
        url: string;
    };
    path: string;
};

export type DeleteEventParams = {
    eventId: string;
    path: string;
};

export type GetAllEventsParams = {
    query: string;
    category: string;
    limit: number;
    page: number;
};

export type GetEventsByUserParams = {
    userId: string;
    limit?: number;
    page: number;
};

export type GetRelatedEventsByCategoryParams = {
    categoryId: string;
    eventId: string;
    limit?: number;
    page: number | string;
};

export type Event = {
    _id: string;
    title: string;
    description: string;
    price: string;
    isFree: boolean;
    imageUrl: string;
    location: string;
    startDateTime: Date;
    endDateTime: Date;
    url: string;
    organizer: {
        _id: string;
        firstName: string;
        lastName: string;
    };
    category: {
        _id: string;
        name: string;
    };
};